0.28: Hash functions and static for loops
Changed testing framework to Catch2.
Code also compiles cleanly with gcc 11.1

+++ lib/include/clu/static_vector.h
@@ -53,7 +53,7 @@ namespace clu
         {
             if constexpr (std::forward_iterator<It>)
             {
-                const auto dist = std::ranges::distance(first, last);
+                const auto dist = static_cast<size_t>(std::ranges::distance(first, last));
                 if (dist > N) size_exceed_capacity();
                 // TODO: remove this MSVC STL issue workaround
                 const auto common = std::views::common(std::ranges::subrange(first, last));
